concentrApp/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
celery-redbeat==2.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-cors-headers==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.2
openapi-codec==1.3.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
exponent_server_sdk
